Finished part of the dash style
